requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
paramiko>=3.0.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.4.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0
jmespath>=1.0.1
debugpy==1.8.1